Fix main branch tests
