+++ packages/client-common/src/utils/sleep.ts
@@ -1,3 +1,7 @@
-export async function sleep(ms: number) {
-  await new Promise((resolve) => setTimeout(resolve, ms))
+export async function sleep(ms: number): Promise<void> {
+  await new Promise((resolve) =>
+    setTimeout(() => {
+      resolve(void 0)
+    }, ms),
+  )
 }
