+++ packages/client-node/__tests__/unit/node_connection_compression.test.ts
@@ -54,7 +54,7 @@ describe('Node.js Connection compression', () => {
       })
 
       const responseBody = 'foobar'
-      emitResponseBody(request, responseBody)
+      await emitResponseBody(request, responseBody)
 
       const queryResult = await selectPromise
       await assertConnQueryResult(queryResult, responseBody)
